inProgress: implementing productManager router
