+++ src/routes/carts.router.js
@@ -1,0 +1,9 @@
+import {Router} from 'express';
+
+const router = Router();
+
+router.get('/', (req, res) => {
+
+});
+
+export default router;
